requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-dotenv
openai
